move studio report modal into modals dir
